project management tool/requirements.txt
rich
python-dateutil
validators
